+++ spec/rubocop/cop/layout/multiline_method_call_indentation_spec.rb
@@ -304,6 +304,15 @@ RSpec.describe RuboCop::Cop::Layout::MultilineMethodCallIndentation do
         RUBY
       end
 
+      context 'target_ruby_version >= 2.5', :ruby25 do
+        it 'accepts key access to hash' do
+          expect_no_offenses(<<~RUBY)
+            hash[key] { 10 / 0 }
+              .fmap { |x| x * 3 }
+          RUBY
+        end
+      end
+
       it 'accepts 3 aligned methods' do
         expect_no_offenses(<<~RUBY)
           a_class.new(severity, location, 'message', 'CopName')